+++ src/main/kotlin/org/olafneumann/regex/generator/output/CSharpCodeGenerator.kt
@@ -1,0 +1,34 @@
+package org.olafneumann.regex.generator.output
+
+import org.olafneumann.regex.generator.regex.RegexCache
+import org.olafneumann.regex.generator.regex.RegexCombiner
+
+internal class CSharpCodeGenerator : SimpleReplacingCodeGenerator(
+    languageName = "C#",
+    highlightLanguage = "csharp",
+    templateCode = """using System;
+using System.Text.RegularExpressions;
+
+public class Sample
+{
+    public static bool useRegex(String input)
+    {
+        const Regex regex = new Regex("%1${'$'}s"%2${'$'}s);
+        return regex.IsMatch(input);
+    }
+}"""
+) {
+
+
+    override fun transformPattern(pattern: String, options: RegexCombiner.Options): String =
+        pattern.replace(RegexCache.get("([\\\\\"])"), "\\\\$1").replace(RegexCache.get("\t"), "\\t")
+
+    override fun generateOptionsCode(options: RegexCombiner.Options) =
+        options.combine(
+            valueForCaseInsensitive = "RegexOptions.IgnoreCase",
+            valueForMultiline = "RegexOptions.Multiline",
+            valueForDotAll = "RegexOptions.Singleline",
+            separator = " | ",
+            prefix = ", "
+        )
+}
